consolidate the `score` mod into the `guild` mod
