+++ src/utils/guild.rs
@@ -1,8 +1,15 @@
 use std::collections::HashMap;
 
 use crate::{
-    models::{db::user::User, transformers::Transformers, user_media_list::MediaListData},
-    utils::{database::establish_connection, score::get_guild_data},
+    models::{
+        db::user::User,
+        transformers::Transformers,
+        user_media_list::{MediaListData, UserMediaList},
+    },
+    utils::{
+        database::establish_connection, queries::FETCH_USER_MEDIA_LIST_DATA,
+        requests::anilist::send_request,
+    },
 };
 
 use serenity::{
@@ -11,6 +18,8 @@ use serenity::{
     model::prelude::{Guild, UserId},
 };
 
+use serde_json::json;
+use tokio::task;
 use tracing::info;
 
 fn get_guild_member_ids(guild: Guild) -> Vec<UserId> {
@@ -39,6 +48,40 @@ pub async fn get_guild_data_for_media<T: Transformers>(
     let mut conn = establish_connection();
     let anilist_users = User::get_users_by_discord_id(guild_members, &mut conn);
     let anilist_users = anilist_users.unwrap();
-    let guild_members_data = get_guild_data(anilist_users, media.get_id(), media.get_type()).await;
+    let guild_members_data =
+        get_guild_anilist_data(anilist_users, media.get_id(), media.get_type()).await;
+    guild_members_data
+}
+
+async fn get_guild_anilist_data(
+    guild_members: Vec<User>,
+    media_id: u32,
+    media_type: String,
+) -> HashMap<i64, MediaListData> {
+    let mut guild_members_data: HashMap<i64, MediaListData> = HashMap::new();
+    for user in guild_members {
+        let body = json!({
+            "query": FETCH_USER_MEDIA_LIST_DATA,
+            "variables": {
+                "userId": user.anilist_id,
+                "type": media_type.to_uppercase(),
+                "mediaId": media_id
+            }
+        });
+        info!("Body: {:#?}", body);
+        let user_media_list_response = task::spawn_blocking(move || send_request(body))
+            .await
+            .unwrap();
+        let user_media_list_response: UserMediaList =
+            serde_json::from_str(&user_media_list_response).unwrap();
+
+        let media_list_data = user_media_list_response.data.unwrap();
+        match media_list_data.media_list {
+            None => continue,
+            Some(data) => {
+                guild_members_data.insert(user.discord_id, data);
+            }
+        };
+    }
     guild_members_data
 }
