+++ src/utils/mod.rs
@@ -7,6 +7,5 @@ pub mod queries;
 pub mod redis;
 pub mod requests;
 pub mod response_fetcher;
-pub mod score;
 pub mod spotify;
 pub mod statics;
